+++ config/di-web.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-use Yiisoft\ErrorHandler\ThrowableHandlerInterface;
-use Yiisoft\ErrorHandler\Handler\ThrowableHandler;
+use Yiisoft\ErrorHandler\Factory\ThrowableResponseFactory;
 use Yiisoft\ErrorHandler\Renderer\HtmlRenderer;
 use Yiisoft\ErrorHandler\ThrowableRendererInterface;
+use Yiisoft\ErrorHandler\ThrowableResponseFactoryInterface;
 
 /**
  * @var array $params
  */
 
 return [
-    ThrowableHandlerInterface::class => ThrowableHandler::class,
     ThrowableRendererInterface::class => HtmlRenderer::class,
+    ThrowableResponseFactoryInterface::class => ThrowableResponseFactory::class,
 ];